Move PermissionFinder to an utility package and restrict it
