+++ src/main/java/org/jenkinsci/plugins/casc/util/PermissionFinder.java
@@ -1,8 +1,9 @@
-package org.jenkinsci.plugins.casc.integrations.globalmatrixauth;
+package org.jenkinsci.plugins.casc.util;
 
 import hudson.security.Permission;
 import hudson.security.PermissionGroup;
-import jenkins.model.Jenkins;
+import org.kohsuke.accmod.Restricted;
+import org.kohsuke.accmod.restrictions.NoExternalUse;
 
 import javax.annotation.CheckForNull;
 import java.util.List;
@@ -11,8 +12,10 @@ import java.util.regex.Pattern;
 
 
 /**
+ * Implements lookup for {@link Permission}s.
  * Created by mads on 2/9/18.
  */
+@Restricted(NoExternalUse.class)
 public class PermissionFinder {
 
     /** For Matrix Auth - Title/Permission **/
@@ -20,7 +23,6 @@ public class PermissionFinder {
 
     /**
      * Attempt to match a given permission to what is defined in the UI.
-     * TODO: Refactor this away when proper permission API is in place for C-as-C
      * @param id String of the form "Title/Permission" (Look in the UI) for a particular permission
      * @return a matched permission
      */
@@ -37,7 +39,7 @@ public class PermissionFinder {
                 }
                 //How do we do this properly, we want to mimic the UI as best as possible. So the logic conclusion is
                 //That when you want admin to be Overall/Administer you put that in. Overall being the group title...
-                //Name being the Permssion you want to set in the matrix.
+                //Name being the Permission you want to set in the matrix.
                 if(pg.title.toString().equals(owner)) {
                     return Permission.fromId(pg.owner.getName()+"."+name);
                 }
